pyproject.toml
[tool.poetry]
name = "summit_project"
version = "0.1.0"
description = ""
authors = ["NeoVic2006 <NeoVic2006@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
Django = "^3.2.6"
djangorestframework = "^3.12.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
